# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for FAS Management System.",
        "title": "FAS Management System API",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "http://localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/schemes": {
            "get": {
                "description": "Retrieve a comprehensive list of all available schemes.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schemes"
                ],
                "summary": "List all schemes",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved list of schemes",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/http.SchemeResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Add a new scheme with the provided details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schemes"
                ],
                "summary": "Create a new scheme",
                "parameters": [
                    {
                        "description": "JSON object containing new scheme details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.CreateSchemeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created scheme",
                        "schema": {
                            "$ref": "#/definitions/http.SchemeResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error occurred",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/schemes/applicant": {
            "get": {
                "description": "Retrieve a list of schemes available for a specific applicant using their unique identifier.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schemes"
                ],
                "summary": "List Applicant Available Schemes",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Applicant ID",
                        "name": "applicant",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved available schemes",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/http.SchemeResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error occurred",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Applicant not found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/schemes/{id}": {
            "get": {
                "description": "Retrieve a scheme using its unique identifier.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schemes"
                ],
                "summary": "Get Scheme by ID",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Scheme ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved scheme",
                        "schema": {
                            "$ref": "#/definitions/http.SchemeResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error occurred",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Scheme not found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Modify the details of an existing scheme using its unique identifier.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schemes"
                ],
                "summary": "Update an existing scheme",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Scheme ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "JSON object with updates to the scheme",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.UpdateSchemeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully updated scheme",
                        "schema": {
                            "$ref": "#/definitions/http.SchemeResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error occurred",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Scheme not found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove a scheme from the system using its unique identifier.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schemes"
                ],
                "summary": "Delete a scheme",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Scheme ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted scheme",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    },
                    "400": {
                        "description": "Validation error occurred",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Scheme not found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "http.CreateSchemeRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "http.ErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    },
                    "example": {
                        "field": "error description"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Error message"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "http.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Success"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "http.SchemeBenefitResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number",
                    "example": 1000000
                },
                "name": {
                    "type": "string",
                    "example": "CDC Vouchers"
                }
            }
        },
        "http.SchemeCriteriaResponse": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "employment_status"
                },
                "value": {
                    "type": "string",
                    "example": "unemployed"
                }
            }
        },
        "http.SchemeResponse": {
            "type": "object",
            "properties": {
                "benefits": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/http.SchemeBenefitResponse"
                    }
                },
                "criteria": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/http.SchemeCriteriaResponse"
                    }
                },
                "id": {
                    "type": "string",
                    "example": "00000000-0000-0000-0000-000000000000"
                },
                "name": {
                    "type": "string",
                    "example": "Retrenchment Assistance Scheme"
                }
            }
        },
...
        "http.UpdateSchemeRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}